requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
cfgrib==0.9.12.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
distlib==0.3.8
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-modeltranslation==0.19.3
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
eccodes==1.7.0
environs==11.0.0
filelock==3.15.1
findlibs==0.0.5
flake8==7.0.0
gunicorn==22.0.0
idna==3.7
kombu==5.3.7
marshmallow==3.21.3
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.1
pathspec==0.12.1
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
pygrib==2.1.5
PyJWT==2.8.0
pyproj==3.6.1
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
